+++ lib/core/utils/navbar_item.dart
@@ -1,0 +1,1 @@
+enum NavbarItem { home, poids, activite, statistiques }