+++ lib/data/states/navigation_state.dart
@@ -1,0 +1,23 @@
+import 'package:freezed_annotation/freezed_annotation.dart';
+import 'package:max_sports/core/utils/navbar_item.dart';
+
+part 'navigation_state.freezed.dart';
+
+@Freezed(makeCollectionsUnmodifiable: false)
+class NavigationState with _$NavigationState {
+  const NavigationState._();
+
+  factory NavigationState.initial() = NavigationStateInitial;
+
+  factory NavigationState.getNavbarItem({
+    required NavbarItem item,
+    required int index,
+  }) = NavigationStateGetNavbarItem;
+
+
+  int get currentIndex => maybeMap(
+    initial: (value) => 0,
+    getNavbarItem: (value) => value.index,
+    orElse: () => 0,
+  );
+}