+++ lib/data/states/navigation_state.freezed.dart
@@ -1,0 +1,334 @@
+// coverage:ignore-file
+// GENERATED CODE - DO NOT MODIFY BY HAND
+// ignore_for_file: type=lint
+// ignore_for_file: unused_element, deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies, annotate_overrides, invalid_annotation_target
+
+part of 'navigation_state.dart';
+
+// **************************************************************************
+// FreezedGenerator
+// **************************************************************************
+
+T _$identity<T>(T value) => value;
+
+final _privateConstructorUsedError = UnsupportedError(
+    'It seems like you constructed your class using `MyClass._()`. This constructor is only meant to be used by freezed and you are not supposed to need it nor use it.\nPlease check the documentation here for more information: https://github.com/rrousselGit/freezed#custom-getters-and-methods');
+
+/// @nodoc
+mixin _$NavigationState {
+  @optionalTypeArgs
+  TResult when<TResult extends Object?>({
+    required TResult Function() initial,
+    required TResult Function(NavbarItem item, int index) getNavbarItem,
+  }) =>
+      throw _privateConstructorUsedError;
+  @optionalTypeArgs
+  TResult? whenOrNull<TResult extends Object?>({
+    TResult Function()? initial,
+    TResult Function(NavbarItem item, int index)? getNavbarItem,
+  }) =>
+      throw _privateConstructorUsedError;
+  @optionalTypeArgs
+  TResult maybeWhen<TResult extends Object?>({
+    TResult Function()? initial,
+    TResult Function(NavbarItem item, int index)? getNavbarItem,
+    required TResult orElse(),
+  }) =>
+      throw _privateConstructorUsedError;
+  @optionalTypeArgs
+  TResult map<TResult extends Object?>({
+    required TResult Function(NavigationStateInitial value) initial,
+    required TResult Function(NavigationStateGetNavbarItem value) getNavbarItem,
+  }) =>
+      throw _privateConstructorUsedError;
+  @optionalTypeArgs
+  TResult? mapOrNull<TResult extends Object?>({
+    TResult Function(NavigationStateInitial value)? initial,
+    TResult Function(NavigationStateGetNavbarItem value)? getNavbarItem,
+  }) =>
+      throw _privateConstructorUsedError;
+  @optionalTypeArgs
+  TResult maybeMap<TResult extends Object?>({
+    TResult Function(NavigationStateInitial value)? initial,
+    TResult Function(NavigationStateGetNavbarItem value)? getNavbarItem,
+    required TResult orElse(),
+  }) =>
+      throw _privateConstructorUsedError;
+}
+
+/// @nodoc
+abstract class $NavigationStateCopyWith<$Res> {
+  factory $NavigationStateCopyWith(
+          NavigationState value, $Res Function(NavigationState) then) =
+      _$NavigationStateCopyWithImpl<$Res>;
+}
+
+/// @nodoc
+class _$NavigationStateCopyWithImpl<$Res>
+    implements $NavigationStateCopyWith<$Res> {
+  _$NavigationStateCopyWithImpl(this._value, this._then);
+
+  final NavigationState _value;
+  // ignore: unused_field
+  final $Res Function(NavigationState) _then;
+}
+
+/// @nodoc
+abstract class _$$NavigationStateInitialCopyWith<$Res> {
+  factory _$$NavigationStateInitialCopyWith(_$NavigationStateInitial value,
+          $Res Function(_$NavigationStateInitial) then) =
+      __$$NavigationStateInitialCopyWithImpl<$Res>;
+}
+
+/// @nodoc
+class __$$NavigationStateInitialCopyWithImpl<$Res>
+    extends _$NavigationStateCopyWithImpl<$Res>
+    implements _$$NavigationStateInitialCopyWith<$Res> {
+  __$$NavigationStateInitialCopyWithImpl(_$NavigationStateInitial _value,
+      $Res Function(_$NavigationStateInitial) _then)
+      : super(_value, (v) => _then(v as _$NavigationStateInitial));
+
+  @override
+  _$NavigationStateInitial get _value =>
+      super._value as _$NavigationStateInitial;
+}
+
+/// @nodoc
+
+class _$NavigationStateInitial extends NavigationStateInitial {
+  _$NavigationStateInitial() : super._();
+
+  @override
+  String toString() {
+    return 'NavigationState.initial()';
+  }
+
+  @override
+  bool operator ==(dynamic other) {
+    return identical(this, other) ||
+        (other.runtimeType == runtimeType && other is _$NavigationStateInitial);
+  }
+
+  @override
+  int get hashCode => runtimeType.hashCode;
+
+  @override
+  @optionalTypeArgs
+  TResult when<TResult extends Object?>({
+    required TResult Function() initial,
+    required TResult Function(NavbarItem item, int index) getNavbarItem,
+  }) {
+    return initial();
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult? whenOrNull<TResult extends Object?>({
+    TResult Function()? initial,
+    TResult Function(NavbarItem item, int index)? getNavbarItem,
+  }) {
+    return initial?.call();
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult maybeWhen<TResult extends Object?>({
+    TResult Function()? initial,
+    TResult Function(NavbarItem item, int index)? getNavbarItem,
+    required TResult orElse(),
+  }) {
+    if (initial != null) {
+      return initial();
+    }
+    return orElse();
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult map<TResult extends Object?>({
+    required TResult Function(NavigationStateInitial value) initial,
+    required TResult Function(NavigationStateGetNavbarItem value) getNavbarItem,
+  }) {
+    return initial(this);
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult? mapOrNull<TResult extends Object?>({
+    TResult Function(NavigationStateInitial value)? initial,
+    TResult Function(NavigationStateGetNavbarItem value)? getNavbarItem,
+  }) {
+    return initial?.call(this);
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult maybeMap<TResult extends Object?>({
+    TResult Function(NavigationStateInitial value)? initial,
+    TResult Function(NavigationStateGetNavbarItem value)? getNavbarItem,
+    required TResult orElse(),
+  }) {
+    if (initial != null) {
+      return initial(this);
+    }
+    return orElse();
+  }
+}
+
+abstract class NavigationStateInitial extends NavigationState {
+  factory NavigationStateInitial() = _$NavigationStateInitial;
+  NavigationStateInitial._() : super._();
+}
+
+/// @nodoc
+abstract class _$$NavigationStateGetNavbarItemCopyWith<$Res> {
+  factory _$$NavigationStateGetNavbarItemCopyWith(
+          _$NavigationStateGetNavbarItem value,
+          $Res Function(_$NavigationStateGetNavbarItem) then) =
+      __$$NavigationStateGetNavbarItemCopyWithImpl<$Res>;
+  $Res call({NavbarItem item, int index});
+}
+
+/// @nodoc
+class __$$NavigationStateGetNavbarItemCopyWithImpl<$Res>
+    extends _$NavigationStateCopyWithImpl<$Res>
+    implements _$$NavigationStateGetNavbarItemCopyWith<$Res> {
+  __$$NavigationStateGetNavbarItemCopyWithImpl(
+      _$NavigationStateGetNavbarItem _value,
+      $Res Function(_$NavigationStateGetNavbarItem) _then)
+      : super(_value, (v) => _then(v as _$NavigationStateGetNavbarItem));
+
+  @override
+  _$NavigationStateGetNavbarItem get _value =>
+      super._value as _$NavigationStateGetNavbarItem;
+
+  @override
+  $Res call({
+    Object? item = freezed,
+    Object? index = freezed,
+  }) {
+    return _then(_$NavigationStateGetNavbarItem(
+      item: item == freezed
+          ? _value.item
+          : item // ignore: cast_nullable_to_non_nullable
+              as NavbarItem,
+      index: index == freezed
+          ? _value.index
+          : index // ignore: cast_nullable_to_non_nullable
+              as int,
+    ));
+  }
+}
+
+/// @nodoc
+
+class _$NavigationStateGetNavbarItem extends NavigationStateGetNavbarItem {
+  _$NavigationStateGetNavbarItem({required this.item, required this.index})
+      : super._();
+
+  @override
+  final NavbarItem item;
+  @override
+  final int index;
+
+  @override
+  String toString() {
+    return 'NavigationState.getNavbarItem(item: $item, index: $index)';
+  }
+
+  @override
+  bool operator ==(dynamic other) {
+    return identical(this, other) ||
+        (other.runtimeType == runtimeType &&
+            other is _$NavigationStateGetNavbarItem &&
+            const DeepCollectionEquality().equals(other.item, item) &&
+            const DeepCollectionEquality().equals(other.index, index));
+  }
+
+  @override
+  int get hashCode => Object.hash(
+      runtimeType,
+      const DeepCollectionEquality().hash(item),
+      const DeepCollectionEquality().hash(index));
+
+  @JsonKey(ignore: true)
+  @override
+  _$$NavigationStateGetNavbarItemCopyWith<_$NavigationStateGetNavbarItem>
+      get copyWith => __$$NavigationStateGetNavbarItemCopyWithImpl<
+          _$NavigationStateGetNavbarItem>(this, _$identity);
+
+  @override
+  @optionalTypeArgs
+  TResult when<TResult extends Object?>({
+    required TResult Function() initial,
+    required TResult Function(NavbarItem item, int index) getNavbarItem,
+  }) {
+    return getNavbarItem(item, index);
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult? whenOrNull<TResult extends Object?>({
+    TResult Function()? initial,
+    TResult Function(NavbarItem item, int index)? getNavbarItem,
+  }) {
+    return getNavbarItem?.call(item, index);
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult maybeWhen<TResult extends Object?>({
+    TResult Function()? initial,
+    TResult Function(NavbarItem item, int index)? getNavbarItem,
+    required TResult orElse(),
+  }) {
+    if (getNavbarItem != null) {
+      return getNavbarItem(item, index);
+    }
+    return orElse();
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult map<TResult extends Object?>({
+    required TResult Function(NavigationStateInitial value) initial,
+    required TResult Function(NavigationStateGetNavbarItem value) getNavbarItem,
+  }) {
+    return getNavbarItem(this);
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult? mapOrNull<TResult extends Object?>({
+    TResult Function(NavigationStateInitial value)? initial,
+    TResult Function(NavigationStateGetNavbarItem value)? getNavbarItem,
+  }) {
+    return getNavbarItem?.call(this);
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult maybeMap<TResult extends Object?>({
+    TResult Function(NavigationStateInitial value)? initial,
+    TResult Function(NavigationStateGetNavbarItem value)? getNavbarItem,
+    required TResult orElse(),
+  }) {
+    if (getNavbarItem != null) {
+      return getNavbarItem(this);
+    }
+    return orElse();
+  }
+}
+
+abstract class NavigationStateGetNavbarItem extends NavigationState {
+  factory NavigationStateGetNavbarItem(
+      {required final NavbarItem item,
+      required final int index}) = _$NavigationStateGetNavbarItem;
+  NavigationStateGetNavbarItem._() : super._();
+
+  NavbarItem get item;
+  int get index;
+  @JsonKey(ignore: true)
+  _$$NavigationStateGetNavbarItemCopyWith<_$NavigationStateGetNavbarItem>
+      get copyWith => throw _privateConstructorUsedError;
+}